Set unix_isatty_job C binding in its own C file
