+++ src/unix/jbuild
@@ -183,6 +183,7 @@
     unix_fstat_job
     unix_fstat_64_job
     unix_utimes_job
+    unix_isatty_job
     not_available
   ))
   (install_c_headers (
