+++ src/unix/lwt_unix_unix.h
@@ -44,35 +44,6 @@
 #include "unix_recv_send_utils.h"
 
 /* +-----------------------------------------------------------------+
-   | JOB: isatty                                                     |
-   +-----------------------------------------------------------------+ */
-
-struct job_isatty {
-    struct lwt_unix_job job;
-    int fd;
-    int result;
-};
-
-static void worker_isatty(struct job_isatty *job)
-{
-    job->result = isatty(job->fd);
-}
-
-static value result_isatty(struct job_isatty *job)
-{
-    value result = Val_bool(job->result);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_isatty_job(value val_fd)
-{
-    LWT_UNIX_INIT_JOB(job, isatty, 0);
-    job->fd = Int_val(val_fd);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-/* +-----------------------------------------------------------------+
    | JOB: opendir                                                    |
    +-----------------------------------------------------------------+ */
